requirements.txt
setuptools==79.0.0
selenium==4.35.0
local==1.0.0
pytest==8.3.2
Appium-Python-Client==4.0.1
pytest-html==4.1.1
pytest-html-reporter==0.2.9
pytest-metadata==3.1.1
pyyaml==6.0.2
pytest-recording==0.13.2
requests==2.32.3
tabulate==0.9.0
cython==3.0.11
pandas==2.2.2
openpyxl==3.1.5
allure-pytest==2.13.5
pytest-dependency==0.6.0
pillow==10.4.0
webdriver-manager==4.0.2
pytest-rerunfailures
pytz
